final draft of foodie service

+++ types/order.types.go
@@ -16,9 +16,9 @@ type PurchaseDetails struct {
 	OrderID    string    `json:"orderId"`
 	Items      []Order   `json:"items"`
 	Products   []Product `json:"products"`
-	TotalPrice int       `json:"totalPrice"`
-	Discount   int       `json:"discount"`
-	FinalPrice int       `json:"finalPrice"`
+	TotalPrice float64      `json:"totalPrice"`
+	Discount   float64       `json:"discount"`
+	FinalPrice float64       `json:"finalPrice"`
 	CouponCode string    `json:"couponCode"`
 	CreatedAt  time.Time `json:"createdAt"`
 	UpdatedAt  time.Time `json:"updatedAt"`
